♻ Move utils to models
